backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
python-jose[cryptography]
passlib[bcrypt]
transformers
python-dotenv